feat: incorporated OakPupilJourneyPreviousLessons BREAKING CHANGE:
