--- src/components/organisms/pupil/OakPreviousLessonsHeading/index.ts
@@ -1,1 +1,0 @@
-export * from "./OakPreviousLessonsHeading";
