+++ src/components/organisms/pupil/OakPupilJourneyPreviousLessons/index.ts
@@ -1,0 +1,1 @@
+export * from "./OakPupilJourneyPreviousLessons";
